factor out potion classes
