+++ src/potions.ts
@@ -1,8 +1,6 @@
 import "core-js/modules/es.object.from-entries";
 import {
   adv1,
-  autosellPrice,
-  availableAmount,
   canAdventure,
   canEquip,
   cliExecute,
@@ -10,7 +8,6 @@ import {
   effectModifier,
   equip,
   getMonsters,
-  haveEffect,
   historicalAge,
   historicalPrice,
   inebrietyLimit,
@@ -25,25 +22,21 @@ import {
   myTurncount,
   numericModifier,
   print,
-  retrievePrice,
   setLocation,
   use,
 } from "kolmafia";
 import {
   $effect,
   $effects,
-  $familiar,
   $item,
   $items,
   $location,
   $slot,
-  clamp,
   ClosedCircuitPayphone,
   CursedMonkeyPaw,
   get,
   getActiveEffects,
   getActiveSongs,
-  getModifier,
   have,
   isSong,
   maxBy,
@@ -58,17 +51,14 @@ import {
   bestShadowRift,
   HIGHLIGHT,
   pillkeeperOpportunityCost,
-  turnsToNC,
   withLocation,
 } from "./lib";
 import { embezzlerCount } from "./embezzler";
-import { usingPurse } from "./outfit";
-import { estimatedGarboTurns } from "./turns";
+import { failedWishes, Potion } from "./potions/potion";
+import { VariableMeatPotion } from "./potions/variable_meat_potion";
 import { globalOptions } from "./config";
 
-export type PotionTier = "embezzler" | "overlap" | "barf" | "ascending";
 const banned = $items`Uncle Greenspan's Bathroom Finance Guide`;
-export const failedWishes: Effect[] = [];
 
 const mutuallyExclusiveList: Effect[][] = [
   $effects`Blue Tongue, Green Tongue, Orange Tongue, Purple Tongue, Red Tongue, Black Tongue`,
@@ -141,286 +131,6 @@ const validPawWishes: Map<Effect, string> = new Map(
       ];
     }),
 );
-
-function retrieveUntradeablePrice(it: Item) {
-  return retrievePrice(it, availableAmount(it) + 1) - autosellPrice(it) * availableAmount(it);
-}
-
-export interface PotionOptions {
-  providesDoubleDuration?: boolean;
-  canDouble?: boolean;
-  effect?: Effect;
-  duration?: number;
-  price?: (historical: boolean) => number;
-  use?: (quantity: number) => boolean;
-  acquire?: (
-    qty: number,
-    item: Item,
-    maxPrice?: number | undefined,
-    throwOnFail?: boolean,
-    maxAggregateCost?: number | undefined,
-    tryRetrievingUntradeable?: boolean,
-  ) => number;
-}
-
-export class Potion {
-  potion: Item;
-  providesDoubleDuration?: boolean;
-  canDouble: boolean;
-  overrideEffect?: Effect;
-  overrideDuration?: number;
-  priceOverride?: (historical: boolean) => number;
-  useOverride?: (quantity: number) => boolean;
-  acquire: (
-    qty: number,
-    item: Item,
-    maxPrice?: number | undefined,
-    throwOnFail?: boolean,
-    maxAggregateCost?: number | undefined,
-    tryRetrievingUntradeable?: boolean,
-  ) => number;
-
-  constructor(potion: Item, options: PotionOptions = {}) {
-    this.potion = potion;
-    this.providesDoubleDuration = options.providesDoubleDuration;
-    this.canDouble = options.canDouble ?? true;
-    this.overrideDuration = options.duration;
-    this.overrideEffect = options.effect;
-    this.priceOverride = options.price;
-    this.useOverride = options.use;
-    this.acquire = options.acquire ?? acquire;
-  }
-
-  doubleDuration(): Potion {
-    if (this.canDouble) {
-      return new Potion(this.potion, {
-        providesDoubleDuration: true,
-        canDouble: this.canDouble,
-        duration: this.overrideDuration,
-        effect: this.overrideEffect,
-        price: this.priceOverride,
-        use: this.useOverride,
-        acquire: this.acquire,
-      });
-    }
-    return this;
-  }
-
-  effect(): Effect {
-    return this.overrideEffect ?? effectModifier(this.potion, "Effect");
-  }
-
-  effectDuration(): number {
-    return (
-      (this.overrideDuration ?? getModifier("Effect Duration", this.potion)) *
-      (this.providesDoubleDuration ? 2 : 1)
-    );
-  }
-
-  meatDrop(): number {
-    return (
-      getModifier("Meat Drop", this.effect()) +
-      2 * (usingPurse() ? getModifier("Smithsness", this.effect()) : 0)
-    );
-  }
-
-  familiarWeight(): number {
-    return getModifier("Familiar Weight", this.effect());
-  }
-
-  bonusMeat(): number {
-    const familiarMultiplier = have($familiar`Robortender`)
-      ? 2
-      : have($familiar`Hobo Monkey`)
-      ? 1.25
-      : 1;
-
-    // Assume base weight of 100 pounds. This is off but close enough.
-    const assumedBaseWeight = 100;
-    // Marginal value of familiar weight in % meat drop.
-    const marginalValue =
-      2 * familiarMultiplier +
-      Math.sqrt(220 * familiarMultiplier) / (2 * Math.sqrt(assumedBaseWeight));
-
-    return this.familiarWeight() * marginalValue + this.meatDrop();
-  }
-
-  static bonusMeat(item: Item): number {
-    return new Potion(item).bonusMeat();
-  }
-
-  gross(embezzlers: number, maxTurns?: number): number {
-    const bonusMeat = this.bonusMeat();
-    const duration = Math.max(this.effectDuration(), maxTurns ?? 0);
-    // Number of embezzlers this will actually be in effect for.
-    const embezzlersApplied = Math.max(
-      Math.min(duration, embezzlers - haveEffect(this.effect())),
-      0,
-    );
-
-    return (bonusMeat / 100) * (baseMeat * duration + 750 * embezzlersApplied);
-  }
-
-  static gross(item: Item, embezzlers: number): number {
-    return new Potion(item).gross(embezzlers);
-  }
-
-  price(historical: boolean): number {
-    if (this.priceOverride) return this.priceOverride(historical);
-    // If asked for historical, and age < 14 days, use historical.
-    // If potion is not tradeable, use retrievePrice instead
-    return this.potion.tradeable
-      ? historical && historicalAge(this.potion) < 14
-        ? historicalPrice(this.potion)
-        : mallPrice(this.potion)
-      : retrieveUntradeablePrice(this.potion);
-  }
-
-  net(embezzlers: number, historical = false): number {
-    return this.gross(embezzlers) - this.price(historical);
-  }
-
-  static net(item: Item, embezzlers: number, historical = false): number {
-    return new Potion(item).net(embezzlers, historical);
-  }
-
-  doublingValue(embezzlers: number, historical = false): number {
-    return Math.min(
-      Math.max(this.doubleDuration().net(embezzlers, historical), 0) -
-        Math.max(this.net(embezzlers, historical), 0),
-      this.price(true),
-    );
-  }
-
-  static doublingValue(item: Item, embezzlers: number, historical = false): number {
-    return new Potion(item).doublingValue(embezzlers, historical);
-  }
-
-  /**
-   * Compute how many times to use this potion to cover the range of turns
-   * @param turns the number of turns to cover
-   * @param allowOverage whether or not to allow the potion to extend past this number of turns
-   * @returns the number of uses required by this potion to cover thatrange
-   */
-  usesToCover(turns: number, allowOverage: boolean): number {
-    if (allowOverage) {
-      return Math.ceil(turns / this.effectDuration());
-    } else {
-      return Math.floor(turns / this.effectDuration());
-    }
-  }
-
-  static usesToCover(item: Item, turns: number, allowOverage: boolean): number {
-    return new Potion(item).usesToCover(turns, allowOverage);
-  }
-
-  /**
-   * Compute how many fewer or more turns we are from the desired turn count based on the input usage
-   * @param turns the number of turns to cover
-   * @param uses the number of uses of hte potion
-   * @returns negative number of the number of turns short, positive number of the number of extra turns
-   */
-  overage(turns: number, uses: number): number {
-    return this.effectDuration() * uses - turns;
-  }
-
-  static overage(item: Item, turns: number, uses: number): number {
-    return new Potion(item).overage(turns, uses);
-  }
-
-  /**
-   * Compute up to 4 possible value thresholds for this potion based on the number of embezzlers to fight at the start of the day
-   * - using it to only cover embezzlers
-   * - using it to cover both barf and embezzlers (this is max 1 use)
-   * - using it to only cover barf
-   * - using it to cover turns in barf and those that would be lost at the end of the day
-   * @param embezzlers The number of embezzlers expected to be fought in a block at the start of the day
-   * @returns
-   */
-  value(
-    embezzlers: number,
-    turns?: number,
-    limit?: number,
-  ): { name: PotionTier; quantity: number; value: number }[] {
-    const startingTurns = haveEffect(this.effect());
-    const ascending = globalOptions.ascend;
-    const totalTurns = turns ?? estimatedGarboTurns();
-    const values: {
-      name: PotionTier;
-      quantity: number;
-      value: number;
-    }[] = [];
-    const limitFunction = limit
-      ? (quantity: number) => clamp(limit - sum(values, ({ quantity }) => quantity), 0, quantity)
-      : (quantity: number) => quantity;
-
-    // compute the value of covering embezzlers
-    const embezzlerTurns = Math.max(0, embezzlers - startingTurns);
-    const embezzlerQuantity = this.usesToCover(embezzlerTurns, false);
-    const embezzlerValue = embezzlerQuantity ? this.gross(embezzlers) : 0;
-
-    values.push({
-      name: "embezzler",
-      quantity: limitFunction(embezzlerQuantity),
-      value: embezzlerValue,
-    });
-
-    // compute the number of embezzlers missed before, and their value (along with barf unless nobarf)
-    const overlapEmbezzlers = -this.overage(embezzlerTurns, embezzlerQuantity);
-
-    if (overlapEmbezzlers > 0) {
-      values.push({
-        name: "overlap",
-        quantity: limitFunction(1),
-        value: this.gross(overlapEmbezzlers, globalOptions.nobarf ? overlapEmbezzlers : undefined),
-      });
-    }
-
-    const embezzlerCoverage =
-      embezzlerQuantity + (overlapEmbezzlers > 0 ? 1 : 0) * this.effectDuration();
-
-    if (!globalOptions.nobarf) {
-      // unless nobarf, compute the value of barf turns
-      // if ascending, break those turns that are not fully covered by a potion into their own value
-      const remainingTurns = Math.max(0, totalTurns - embezzlerCoverage - startingTurns);
-
-      const barfQuantity = this.usesToCover(remainingTurns, !ascending);
-      values.push({ name: "barf", quantity: limitFunction(barfQuantity), value: this.gross(0) });
-
-      if (ascending && this.overage(remainingTurns, barfQuantity) < 0) {
-        const ascendingTurns = Math.max(0, remainingTurns - barfQuantity * this.effectDuration());
-        values.push({
-          name: "ascending",
-          quantity: limitFunction(1),
-          value: this.gross(0, ascendingTurns),
-        });
-      }
-    }
-
-    return values.filter((tier) => tier.quantity > 0);
-  }
-
-  private _use(quantity: number): boolean {
-    if (this.useOverride) {
-      return this.useOverride(quantity);
-    } else if (itemType(this.potion) === "potion") {
-      return use(quantity, this.potion);
-    } else {
-      // must provide an override for non potions, otherwise they won't use
-      return false;
-    }
-  }
-
-  use(quantity: number): boolean {
-    const effectTurns = haveEffect(this.effect());
-    const result = this._use(quantity);
-    // If we tried wishing but failed, no longer try this wish in the future
-    if (this.potion === $item`pocket wish` && haveEffect(this.effect()) <= effectTurns) {
-      failedWishes.push(this.effect());
-    }
-    return result;
-  }
-}
 
 function useAsValuable(potion: Potion, embezzlers: number, embezzlersOnly: boolean): number {
   const value = potion.value(embezzlers);
@@ -702,133 +412,6 @@ export function bathroomFinance(embezzlers: number): void {
       print(`Using ${greenspan}!`, HIGHLIGHT);
       use(greenspan);
     }
-  }
-}
-
-function triangleNumber(b: number, a = 0) {
-  return 0.5 * (b * (b + 1) - a * (a + 1));
-}
-
-class VariableMeatPotion {
-  potion: Item;
-  effect: Effect;
-  duration: number;
-  softcap: number; // Number of turns to cap out variable bonus
-  meatBonusPerTurn: number; // meat% bonus per turn
-  cappedMeatBonus: number;
-
-  constructor(
-    potion: Item,
-    softcap: number,
-    meatBonusPerTurn: number,
-    duration?: number,
-    effect?: Effect,
-  ) {
-    this.potion = potion;
-    this.effect = effect ?? effectModifier(potion, "Effect");
-    this.duration = duration ?? numericModifier(potion, "Effect Duration");
-    this.softcap = softcap;
-    this.meatBonusPerTurn = meatBonusPerTurn;
-    this.cappedMeatBonus = softcap * meatBonusPerTurn;
-  }
-
-  use(quantity: number): boolean {
-    acquire(
-      quantity,
-      this.potion,
-      (1.2 * retrievePrice(this.potion, quantity)) / quantity,
-      false,
-      2000000,
-    );
-    if (availableAmount(this.potion) < quantity) return false;
-    return use(quantity, this.potion);
-  }
-
-  price(historical: boolean): number {
-    // If asked for historical, and age < 14 days, use historical.
-    // If potion is not tradeable, use retrievePrice instead
-    return this.potion.tradeable
-      ? historical && historicalAge(this.potion) < 14
-        ? historicalPrice(this.potion)
-        : mallPrice(this.potion)
-      : retrieveUntradeablePrice(this.potion);
-  }
-
-  getOptimalNumberToUse(yachtzees: number, embezzlers: number): number {
-    const barfTurns = Math.max(0, estimatedGarboTurns() - yachtzees - embezzlers);
-
-    const potionAmountsToConsider: number[] = [];
-    const considerSoftcap = [0, this.softcap];
-    const considerEmbezzlers = embezzlers > 0 ? [0, embezzlers] : [0];
-    for (const fn of [Math.floor, Math.ceil]) {
-      for (const sc of considerSoftcap) {
-        for (const em of considerEmbezzlers) {
-          const considerBarfTurns = em === embezzlers && barfTurns > 0 ? [0, barfTurns] : [0];
-          for (const bt of considerBarfTurns) {
-            const potionAmount = fn((yachtzees + em + bt + sc) / this.duration);
-            if (!potionAmountsToConsider.includes(potionAmount)) {
-              potionAmountsToConsider.push(potionAmount);
-            }
-          }
-        }
-      }
-    }
-
-    const profitsFromPotions = potionAmountsToConsider.map((quantity) => ({
-      quantity,
-      value: this.valueNPotions(quantity, yachtzees, embezzlers, barfTurns),
-    }));
-    const bestOption = maxBy(profitsFromPotions, "value");
-
-    if (bestOption.value > 0) {
-      print(
-        `Expected to profit ${bestOption.value.toFixed(2)} from ${bestOption.quantity} ${
-          this.potion.plural
-        }`,
-        "blue",
-      );
-      const ascendingOverlap =
-        globalOptions.ascend || globalOptions.nobarf ? 0 : this.softcap / this.duration;
-      const potionsToUse =
-        bestOption.quantity +
-        ascendingOverlap -
-        Math.floor(haveEffect(this.effect) / this.duration);
-
-      return Math.max(potionsToUse, 0);
-    }
-    return 0;
-  }
-
-  valueNPotions(n: number, yachtzees: number, embezzlers: number, barfTurns: number): number {
-    const yachtzeeValue = 2000;
-    const embezzlerValue = baseMeat + 750;
-    const barfValue = (baseMeat * turnsToNC) / 30;
-
-    const totalCosts = retrievePrice(this.potion, n);
-    const totalDuration = n * this.duration;
-    let cappedDuration = Math.max(0, totalDuration - this.softcap + 1);
-    let decayDuration = Math.min(totalDuration, this.softcap - 1);
-    let totalValue = 0;
-    const turnTypes = [
-      [yachtzees, yachtzeeValue],
-      [embezzlers, embezzlerValue],
-      [barfTurns, barfValue],
-    ];
-
-    for (const [turns, value] of turnTypes) {
-      const cappedTurns = Math.min(cappedDuration, turns);
-      const decayTurns = Math.min(decayDuration, turns - cappedTurns);
-      totalValue +=
-        (value *
-          (cappedTurns * this.cappedMeatBonus +
-            triangleNumber(decayDuration, decayDuration - decayTurns) * this.meatBonusPerTurn)) /
-        100;
-      cappedDuration -= cappedTurns;
-      decayDuration -= decayTurns;
-      if (decayDuration === 0) break;
-    }
-
-    return totalValue - totalCosts;
   }
 }
